Perspective and orthographic both kinda work now sorta
Perspective and orthographic rendering aren't handling device z
coordinate in a consistent fashion, which is probably to be expected.

I think I should just give up on an infinite far plane, and switch to
using reverse-z w/ a 32 bit floating point depth attachment.  This
would get better precision anyway, and would probably clear up the
rendering artifacts I'm seeing in ortho.

This would necessitate using a render target for rendering the scene
and then having a pass that draws it into the back buffer, but I'll
need to do that anyway sooner or later.